convert from contact to ci
changing from a simple "contact" to a "configuration item" or "ci". using a "ci" gives a lot more flexibility on monitoring system side to define policies based on the configuration item. this commit also allows you to disable integration with cork and indeed the integration is disabled by default. a couple more changes are on the way to the registration page and then a new version will be tagged.